# wafaa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "API pour gérer les artistes, albums et chansons",
        "contact": {
            "email": "admin@musicbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Liste des albums",
                "operationId": "f8fe3c749d4ce165cebbbc29f4fd774b",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filtrer par année",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "artist_id",
                        "in": "query",
                        "description": "Filtrer par ID d'artiste",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des albums"
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Créer un nouvel album",
                "operationId": "2a22c76c013f500774befbe4b3a7bba1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "year",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Divide"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2017
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album créé avec succès"
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Afficher un album",
                "operationId": "7abaec732d77679e785a57170ce30f79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album trouvé"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Mettre à jour un album",
                "operationId": "9796938e22ad1d24d41db760a148fb45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "No.6 Collaborations Project"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2019
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album mis à jour"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Supprimer un album",
                "operationId": "716f6669369d76aedc1aee8be959ecab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album supprimé"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            }
        },
        "/api/albums/{id}/songs": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister les chansons d'un album",
                "operationId": "ccfbacbce34fddf8cb95ecf99a7de0ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chansons de l'album"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            }
        },
...
    },
...
}